+++ mayan/apps/appearance/apps.py
@@ -4,6 +4,7 @@ from mayan.apps.common.apps import MayanAppConfig
 
 
 class AppearanceApp(MayanAppConfig):
+    app_url = 'appearance'
     has_tests = True
     name = 'mayan.apps.appearance'
     verbose_name = _('Appearance')