+++ mayan/apps/common/apps.py
@@ -212,11 +212,11 @@ class CommonApp(MayanAppConfig):
 
             if os.path.exists(settings.MEDIA_ROOT) and setting_production_error_logging.value:
                 logging_configuration['handlers']['logfile'] = {
-                    'backupCount': 3,
+                    'backupCount': 5,
                     'class': 'logging.handlers.RotatingFileHandler',
                     'filename': setting_production_error_log_path.value,
                     'formatter': 'logfile',
-                    'maxBytes': 1024,
+                    'maxBytes': 65535,
                 }
 
             logging.config.dictConfig(logging_configuration)