+++ HISTORY.rst
@@ -4,7 +4,8 @@
 - Fix layer transformation selection view.
 - Improve permission checking of the layer transformation
   selection view.
-- Make tag widget clickable.
+- Make document tag widget clickable.
+- Make document cabinet widget clickable.
 
 3.4.2 (2020-04-02)
 ==================