+++ HISTORY.rst
@@ -6,6 +6,8 @@
   selection view.
 - Make document tag widget clickable.
 - Make document cabinet widget clickable.
+- Apply the ``DOCUMENTS_LIST_THUMBNAIL_WIDTH`` setting value to
+  document pages and document version thumbnails too.
 
 3.4.2 (2020-04-02)
 ==================